+++ lib/ts/recipeImplementation.ts
@@ -1,17 +1,25 @@
-import { RecipeInterface, NormalisedInputType } from "./types";
+import { RecipeInterface, EventHandler, RecipePreAPIHookFunction, RecipePostAPIHookFunction } from "./types";
 import AuthHttpRequest, { FrontToken, getIdRefreshToken } from "./fetch";
 import { interceptorFunctionRequestFulfilled, responseInterceptor, responseErrorInterceptor } from "./axios";
 import { supported_fdi } from "./version";
 import { logDebugMessage } from "./logger";
 
-export default function RecipeImplementation(): RecipeInterface {
+export default function RecipeImplementation(recipeImplInput: {
+    preAPIHook: RecipePreAPIHookFunction;
+    postAPIHook: RecipePostAPIHookFunction;
+    onHandleEvent: EventHandler;
+    sessionExpiredStatusCode: number;
+}): RecipeInterface {
     return {
-        addFetchInterceptorsAndReturnModifiedFetch: function(originalFetch: any, _: NormalisedInputType): typeof fetch {
+        addFetchInterceptorsAndReturnModifiedFetch: function(input: {
+            originalFetch: any;
+            userContext: any;
+        }): typeof fetch {
             logDebugMessage("addFetchInterceptorsAndReturnModifiedFetch: called");
             return async function(url: RequestInfo, config?: RequestInit): Promise<Response> {
                 return await AuthHttpRequest.doRequest(
                     (config?: RequestInit) => {
-                        return originalFetch(typeof url === "string" ? url : url.clone(), {
+                        return input.originalFetch(typeof url === "string" ? url : url.clone(), {
                             ...config
                         });
                     },
@@ -20,10 +28,10 @@ export default function RecipeImplementation(): RecipeInterface {
                 );
             };
         },
-        addAxiosInterceptors: function(axiosInstance: any, _: NormalisedInputType): void {
+        addAxiosInterceptors: function(input: { axiosInstance: any; userContext: any }): void {
             logDebugMessage("addAxiosInterceptors: called");
             // we first check if this axiosInstance already has our interceptors.
-            let requestInterceptors = axiosInstance.interceptors.request;
+            let requestInterceptors = input.axiosInstance.interceptors.request;
             for (let i = 0; i < requestInterceptors.handlers.length; i++) {
                 if (requestInterceptors.handlers[i].fulfilled === interceptorFunctionRequestFulfilled) {
                     logDebugMessage("addAxiosInterceptors: not adding because already added on this instance");
@@ -31,17 +39,19 @@ export default function RecipeImplementation(): RecipeInterface {
                 }
             }
             // Add a request interceptor
-            axiosInstance.interceptors.request.use(interceptorFunctionRequestFulfilled, async function(error: any) {
+            input.axiosInstance.interceptors.request.use(interceptorFunctionRequestFulfilled, async function(
+                error: any
+            ) {
                 throw error;
             });
 
             // Add a response interceptor
-            axiosInstance.interceptors.response.use(
-                responseInterceptor(axiosInstance),
-                responseErrorInterceptor(axiosInstance)
+            input.axiosInstance.interceptors.response.use(
+                responseInterceptor(input.axiosInstance),
+                responseErrorInterceptor(input.axiosInstance)
             );
         },
-        getUserId: async function(_: NormalisedInputType): Promise<string> {
+        getUserId: async function(_: { userContext: any }): Promise<string> {
             logDebugMessage("getUserId: called");
             let tokenInfo = await FrontToken.getTokenInfo();
             if (tokenInfo === undefined) {
@@ -50,7 +60,7 @@ export default function RecipeImplementation(): RecipeInterface {
             logDebugMessage("getUserId: returning: " + tokenInfo.uid);
             return tokenInfo.uid;
         },
-        getAccessTokenPayloadSecurely: async function(config: NormalisedInputType): Promise<any> {
+        getAccessTokenPayloadSecurely: async function(input: { userContext: any }): Promise<any> {
             logDebugMessage("getAccessTokenPayloadSecurely: called");
             let tokenInfo = await FrontToken.getTokenInfo();
             if (tokenInfo === undefined) {
@@ -61,7 +71,9 @@ export default function RecipeImplementation(): RecipeInterface {
                 logDebugMessage("getAccessTokenPayloadSecurely: access token expired. Refreshing session");
                 let retry = await AuthHttpRequest.attemptRefreshingSession();
                 if (retry) {
-                    return await this.getAccessTokenPayloadSecurely(config);
+                    return await this.getAccessTokenPayloadSecurely({
+                        userContext: input.userContext
+                    });
                 } else {
                     throw new Error("Could not refresh session");
                 }
@@ -69,23 +81,24 @@ export default function RecipeImplementation(): RecipeInterface {
             logDebugMessage("getAccessTokenPayloadSecurely: returning: " + JSON.stringify(tokenInfo.up));
             return tokenInfo.up;
         },
-        doesSessionExist: async function(_: NormalisedInputType): Promise<boolean> {
+        doesSessionExist: async function(_: { userContext: any }): Promise<boolean> {
             logDebugMessage("doesSessionExist: called");
             return (await getIdRefreshToken(true)).status === "EXISTS";
         },
-        signOut: async function(config: NormalisedInputType): Promise<void> {
+        signOut: async function(input: { userContext: any }): Promise<void> {
             logDebugMessage("signOut: called");
-            if (!(await this.doesSessionExist(config))) {
+            if (!(await this.doesSessionExist(input))) {
                 logDebugMessage("signOut: existing early because session does not exist");
                 logDebugMessage("signOut: firing SIGN_OUT event");
-                config.onHandleEvent({
-                    action: "SIGN_OUT"
+                recipeImplInput.onHandleEvent({
+                    action: "SIGN_OUT",
+                    userContext: input.userContext
                 });
                 return;
             }
 
             logDebugMessage("signOut: Calling refresh pre API hook");
-            let preAPIResult = await config.preAPIHook({
+            let preAPIResult = await recipeImplInput.preAPIHook({
                 action: "SIGN_OUT",
                 requestInit: {
                     method: "post",
@@ -94,15 +107,27 @@ export default function RecipeImplementation(): RecipeInterface {
                         rid: AuthHttpRequest.rid
                     }
                 },
-                url: AuthHttpRequest.signOutUrl
+                url: AuthHttpRequest.signOutUrl,
+                userContext: input.userContext
             });
 
             logDebugMessage("signOut: Calling API");
             let resp = await fetch(preAPIResult.url, preAPIResult.requestInit);
             logDebugMessage("signOut: API ended");
 
+            await recipeImplInput.postAPIHook({
+                action: "SIGN_OUT",
+                requestInit: preAPIResult.requestInit,
+                url: preAPIResult.url,
+                fetchResponse: resp.clone(),
+                userContext: input.userContext
+            });
+
+            logDebugMessage("signOut: API ended");
+
             logDebugMessage("signOut: API responded with status code: " + resp.status);
-            if (resp.status === config.sessionExpiredStatusCode) {
+
+            if (resp.status === recipeImplInput.sessionExpiredStatusCode) {
                 // refresh must have already sent session expiry event
                 return;
             }