+++ lib/ts/utils/windowHandler/types.ts
@@ -12,6 +12,22 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
+export type StorageHandler = {
+    key: (index: number) => Promise<string | null>;
+    getItem: (key: string) => Promise<string | null>;
+    clear: () => Promise<void>;
+    removeItem: (key: string) => Promise<void>;
+    setItem: (key: string, value: string) => Promise<void>;
+    /**
+     * Sync versions of the storage functions
+     */
+    keySync: (index: number) => string | null;
+    getItemSync: (key: string) => string | null;
+    clearSync: () => void;
+    removeItemSync: (key: string) => void;
+    setItemSync: (key: string, value: string) => void;
+};
+
 export type WindowHandlerInterface = {
     history: {
         replaceState: (data: any, unused: string, url?: string | null) => void;
@@ -28,8 +44,8 @@ export type WindowHandlerInterface = {
         getHostName: () => string;
     };
     getDocument: () => Document;
-    getLocalStorage: () => Storage;
-    getSessionStorage: () => Storage;
+    localStorage: StorageHandler;
+    sessionStorage: StorageHandler;
 };
 
 export type WindowHandlerInput = (original: WindowHandlerInterface) => WindowHandlerInterface;